The problems of pulling yourself up by your bootstraps!
The Processor subsystems now know the following about the system they
will be a part of:

* Each core has a workspace of type core_workspace
* That workspace includes a structure called svc_stack
* All cores share a workspace of type shared_workspace

The structs are defined in a header file called system_workspaces.h,
in whichever system's directory you're building.

The locations of the structures in memory are the same for all cores,
and determined on the gcc command line. As is the amount of RAM to
assume on startup and the locations of the translation tables accessible
to kernel memory. (Currently all in the top MiB.)

When the higher level code is entered (in high memory, caches and MMU
enabled, and the stack pointer pointing at the top of workspace.svc_stack),
there are still no processor vectors established, although they will be
located at 0xffff0000 (SCTLR.V == 1).