The workspaces are now cleared before boot_with_stack is called
This means that any locks in the shared memory can be assumed to be
unclaimed, and the core-specific workspace is clear before it is used
for the svc stack.

The core_claim_lock routine returns true if the lock already contained
the value that indicates that the core claimed the lock. If it's initialised
to 1, that means only core 0 will pass through the claim function and
return true, and the other cores will have to wait for the lock to be
released again (without the normal "if (!reclaimed) ...").