Blinking again
The Tests/OSTasks/blink.c successfully blinks the LEDs from one OSTask,
probably outputs to the UART (it works on qemu), from another, and
Sleep appears to be working properly.

The next step is to run the separate tasks in separate slots, using the
OSTask_Spawn SWI. They should be able to allocate slot-local mappings to
their various devices and not interfere with each other.

Next, handle different interrupts, not just "One Flavor"!

That will allow for another spawned task to manage the GPU mailboxes, etc.
and open displays.