Preparation for legacy SWIs
Tests/OSTasks builds and runs on qemu and real hardware, but one
time on real hardware the LEDs both came on and didn't blink. (I tried
several more times, but it always appeared to work.)

The legacy environment will consist of at least the following:

* A 1MiB legacy svc stack (a smaller one might be possible)
* Vectors
* Callbacks
* Transient callbacks
* Service call handlers
* Legacy kernel workspace
  - relocated, if at all possible away from 0xffff0000
  - made user accessible, so that legacy code can be run in that
    mode during the transition period?

etc.

At most one task at a time can be the chosen legacy task; the modified
Wimp will decide which one (between calls to Wimp_Poll, or Wimp_SendMessage,
etc.).

That's the idea at the moment, anyway.

All other SWIs will never enable interrupts, ones that would need to will
hand off control of the calling task to another task (running usr32 code)
via a task queue. (Task queues transfer the task handle, the swi number,
and the core number of each task that gets placed into them, relinquishing
control to the owner of the queue.)

Modules will continue to provide commands and up to 64 SWIs. To interact
with the C kernel, they will register a block of 64 pointers to either
functions or queues. If that fails, they're using an old kernel and can
run in the old manner.

The "UtilityModule" (Kernel) will have a block of 128 functions or queues,
many copied from the JTABLE in an existing ROM image.