Adding legacy RISC OS code support
Looking at the kernel code again, it seems you can't spit without hitting
a routine enabling interrupts! Even quite short ones.

I have the idea that routines can be allowed to run if the svc stack is the
legacy stack, or to claim that stack if nobody else is using it.

If another task is using the legacy stack, the call gets queued and a server
task removes them one at a time and gets the original task to run CallASWIR12
with the original registers in its own slot.

Can I ensure that the task never gets its CallASWIR12 task queued again,
causing a potential deadlock?

Find out next week...