Started to add module functionality
The new kernel has to take control of all the module initialisation
and execution, as well as anything to do with memory management.

* OS_Module
* OS_ReadDynamicArea
* OS_DynamicArea
* OS_ChangeDynamicArea
* OS_Memory (with a view to elimination, far too much implmentation exposed)
* OS_Heap (possibly, possibly not)

New modules have a SWI to register how they deal with each of their (up to)
64 SWIs; each one may execute immediately (without enabling interrupts) or
queue the request to be dealt with by a user mode task.

The Legacy subsystem already makes use of the building blocks that make that
possible.