Multiple corrections for multi-processing
The mpsafe... functions weren't. I think they are now.

There are facilities to run a (usr mode) task on a specific core, but
that shouldn't be used any more than absolutely necessary.

Ditto the RunPrivileged SWI which is needed to access system registers
for timers in the QA7 module for QEMU.

The current problem is interrupt handling.

Specifically, the QA7 element is not an interrupt controller that can
disable an incoming interrupt until we're ready for it.

Needs more thought, but an interim check in seemed like a good idea.

Still using Tests/Legacy/build as the main build script.