Display is working
What needs fixing is the passing of a physical address to a SWI;
any user task could do that with any physical address, which is not
a good idea, to say the least.

Instead, I should create a pipe over the memory in question (using my
task's virtual address for it), pass the pipe to the SWI, at which point
the server may read it as a block.

Perhaps a new PipeOp, PipeOp_Reverse will allow the server to write to
the memory and the client to read it? A new creation routine could be
called PipeOp_CreateConversation?

For security, the pipe would need a reversible flag...

Incidentally, the QA7 module now uses the local timer, which re-loads itself
on reaching zero, which means it shouldn't drift. Why I didn't think it was
supported under QEMU, I don't know.