Fixed a bug when creating a shared global area
Specifically, the legacy svc stack was causing problems because the
global map of the table wasn't being updated.

The temporary DEBUG__BREAKME_STARTUP_CORE_NON_ZERO debug code tends
to show up MMU problems with global memory areas because it ensures
that the initialisation code is run on a core other than core 0.