Re-organised Legacy code a bit
Fixed a problem where interrupts were corrupting the user mode
banked registers for the legacy manage_legacy_stack task (only)
because default behaviour is to store those registers for when
the interrupt returns.

That's not a problem for any native code because it will not
enable interrupts in privileged modes.