Changes for multiple legacy stack owners
Only one active OSTask can be using the legacy SVC stack at a time, but
there can be several that have become blocked during the execution of a
legacy SWI.

Access to the stack is controlled by a trivial user task that is blocked
while SWIs are being run.

There is a bug where a task is given the wrong banked register values.